% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_A2}
\alias{df_A2}
\title{Simulated Clinical Data for mapping}
\format{
A data.frame containing simulated clinical features.
}
\source{
Simulated clinical data for AFRIN package.
}
\usage{
df_A2
}
\description{
A simulated clinical dataset used to demonstrate simple mapping in the AFRIN package.
}
\keyword{datasets}
